feat: add lfo
The changes in this commit add support for Low Frequency Oscillators (LFOs) to the audio node context and workspace components. This will allow for the creation and manipulation of LFOs in the application, enabling more complex and dynamic audio synthesis.

Here's a breakdown of the changes:

1. `AudioNodeContext.tsx`:
   - Added a new type `'lfo'` to the `ModuleNode` union type, which represents an LFO module.
   - Added `LfoOptions` type to define the construction properties of an LFO module.

2. `Workspace.tsx`:
  - Updated the ModuleType union type to include the new 'lfo' type.
  - Updated the ModuleData union type to include data for LFO modules.

3. `Toolbar.tsx`:
   - Added a new button to the toolbar for creating LFO modules.

4. `Oscillator.tsx`:
   - Updated the `Oscillator` component to handle LFO modules as well as oscillator modules.

Testing:
To test these changes, you can create a new LFO module by clicking the LFO button in the toolbar. You should be able to see the LFO module appear in the workspace, and you should be able to manipulate its properties using the controls in the module.